finalizando o cadastro de cliente
