+++ src/Pages/Clientes/AtualizarCliente/index.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import styled from "styled-components";
 import Cabecalho from "../../../Components/Cabecalho";
-import FormsCliente from "../../../Components/Forms/formsCliente";
 
 const Main = styled.main`
     width: 100%;
@@ -18,7 +17,7 @@ function AtualizaCliente(){return(
     <Cabecalho />
     <Main>
         <ExibeTitulo id="exibe-titulo" className="exibe-titulo">Atualizar Cliente</ExibeTitulo>
-        <FormsCliente />
+
     </Main>
     </>
 )};
